+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 CHANGELOG
 =====
 
-# 201707xx
+# 20170721
+* Added support for message_type for debugging method (0 = info, 1 = warning, 2 = error)
+* Added more debug output
+* Updated several cheatsheets
+* Search now shows a score for each cheatcheet which contains matches (#1)
+
+# 20170720
 * Renamed project from CHEAT to CHEATER
 * Added debugging function to utils.py
 * Search now features 'searchstring'
@@ -9,3 +15,4 @@ CHANGELOG
 * Search now features amount of matching lines
 * Search is now no longer case-specific
 * Listing now shows number of cheatsheets
+* Added debug switch